fix: Merge common functions
