+++ package.json
@@ -40,6 +40,7 @@
     "@nuxt/module-builder": "^0.2.1",
     "@nuxt/schema": "^3.1.1",
     "@nuxtjs/eslint-config-typescript": "^12.0.0",
+    "@types/cookie": "^0.5.1",
     "eslint": "^8.33.0",
     "nuxt": "^3.1.1",
     "ofetch": "^1.0.0"
